Renamed scenario_config_parser to scenario_parser
